test_cp_inode_invalidate test: Fix most comments.
